refactoring to graduation folder
